+++ SRC/handler/DataFileStream.cpp
@@ -310,9 +310,9 @@ DataFileStream::write(Vector &data)
   // write data
   if (doCSV == 0) {
     for (int i=0; i<maxCount+1; i++) {
-      int fileID = printMapping(0,i);
-      int startLoc = printMapping(1,i);
-      int numData = printMapping(2,i);
+      int fileID = (int)printMapping(0,i);
+      int startLoc = (int)printMapping(1,i);
+      int numData = (int)printMapping(2,i);
       double *data = theData[fileID];
       for (int j=0; j<numData; j++)
 	theFile << data[startLoc++] << " ";
@@ -320,9 +320,9 @@ DataFileStream::write(Vector &data)
     theFile << "\n";
   } else {
     for (int i=0; i<maxCount+1; i++) {
-      int fileID = printMapping(0,i);
-      int startLoc = printMapping(1,i);
-      int numData = printMapping(2,i);
+      int fileID = (int)printMapping(0,i);
+      int startLoc = (int)printMapping(1,i);
+      int numData = (int)printMapping(2,i);
       double *data = theData[fileID];
       int nM1 = numData-1;
       for (int j=0; j<numData; j++)
